requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langgraph
slowapi
python-dotenv
requests
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
